This PR contains a few simple optimizations that drastically improve the performance of the FRR northbound layer.

With these optimizations, the northbound should be able to load large configurations in a few seconds instead of minutes. Also, the transaction processing now scales linearly according to the size of the candidate being committed.

The table below shows some numbers I got during my tests (`time(1)` was used to measure how much time ripd takes to load large configurations using either the CLI or the gRPC plugin):

|                             | CLI     | gRPC/JSON |
| ----------------------------| ------- | --------- |
| **100 thousand RIP routes** | 1.161s  | 0.574s    |
| **1 million RIP routes**    | 11.885s | 5.817s    |

Other commands like `route-map` take slightly slower to process. This depends on the complexity of the associated YANG module and how many validations libyang needs to perform (and how many default nodes it needs to create).

This PR still doesn't solve the performance issues completely when `--tcli` is not used (i.e. when the northbound creates a separate transaction for every configuration command). This can only be solved by optimizing the libyang `lyd_validate()` function, and work is being done to get that done, along with other libyang general performance enhancements.

No regressions were detected using ANVL. Please refer to the individual commit messages for more details.